+++ packages/ts-ioc-container/__tests__/hooks/hook.spec.ts
@@ -17,6 +17,7 @@ import {
   register,
   Registration as R,
   sequential,
+  by,
 } from '../../lib';
 import { perform, runSequential, runSync } from './runners';
 
@@ -64,7 +65,7 @@ describe('hooks', () => {
       @hook('syncBefore', (ctx) => {
         ctx.invokeMethod();
       })
-      start(@inject(arg(0)) firstArg: string, @inject('suffix') suffix: string, runtimeArg: string) {
+      start(@inject(arg(0)) firstArg: string, @inject(by('suffix')) suffix: string, runtimeArg: string) {
         this.receivedArgs = [firstArg, suffix, runtimeArg];
       }
     }
@@ -90,7 +91,7 @@ describe('hooks', () => {
       @hook('syncBefore', (ctx) => {
         ctx.invokeMethod();
       })
-      start(@inject(arg(0)) firstArg: string, @inject('suffix') suffix: string) {
+      start(@inject(arg(0)) firstArg: string, @inject(by('suffix')) suffix: string) {
         this.receivedArgs = [firstArg, suffix];
       }
     }
@@ -138,13 +139,13 @@ describe('hooks', () => {
       isStarted = false;
 
       @hook('onStart', executeAsync)
-      async initialize(@inject('TimeToSleep') timeToSleep: number) {
+      async initialize(@inject(by('TimeToSleep')) timeToSleep: number) {
         await sleep(timeToSleep);
         this.isStarted = true;
       }
 
       @hook('onStart', executeAsync)
-      async dispose(@inject('TimeToSleep') timeToSleep: number) {
+      async dispose(@inject(by('TimeToSleep')) timeToSleep: number) {
         await sleep(timeToSleep);
         this.isStarted = false;
       }
@@ -454,7 +455,7 @@ describe('hooks', () => {
     }
 
     class App {
-      constructor(@inject(PluginToken.lazy()) private readonly plugins: Plugin[]) {}
+      constructor(@inject(by(PluginToken.lazy())) private readonly plugins: Plugin[]) {}
 
       runPlugins(scope: Container) {
         this.plugins.forEach((plugin) => runOnPluginStart(plugin, { scope }));